+++ highs/ipm/hipo/auxiliary/Auxiliary.h
@@ -35,8 +35,8 @@ void dfsPostorder(Int node, Int& start, std::vector<Int>& head,
 void processEdge(Int j, Int i, const std::vector<Int>& first,
                  std::vector<Int>& maxfirst, std::vector<Int>& delta,
                  std::vector<Int>& prevleaf, std::vector<Int>& ancestor);
-double getDiagStart(Int n, Int k, Int nb, Int n_blocks, std::vector<Int>& start,
-                    bool triang = false);
+int64_t getDiagStart(Int n, Int k, Int nb, Int n_blocks,
+                     std::vector<Int>& start, bool triang = false);
 void firstDescendant(const std::vector<Int>& parent, std::vector<Int>& first);
 
 template <typename T>
